+++ src/adapters/supabase/types/index.ts
@@ -1,1 +1,1 @@
-export * from "./database.types";
+export * from "./database";
