+++ src/assets/svgs/index.ts
@@ -1,4 +1,4 @@
-export * from "./_commit";
+export * from "./commit";
 export * from "./opened-issue";
 export * from "./closed-issue";
 export * from "./opened-pull";
